Refactor Whisper driver by removing `batch.rs`, relocating batching logic to `context.rs` for centralized processing, and modularizing the driver into separate `context` and `driver_impl` modules. Update imports, streamline structure, and enhance clarity.
